requirements.txt
opencv-python==4.5.3.56
numpy==1.21.2
matplotlib==3.4.3
tensorflow==2.6.0
pillow==8.3.2
scikit-learn==0.24.2